Add deleteTodo test
