+++ src/deleteTodo.js
@@ -1,0 +1,10 @@
+import { IndexArrangement } from './todo';
+
+const deleteTodo = (arr) => {
+  if (arr.length > 0) {
+    const newArr = arr.filter((todo) => todo.completed === false);
+    IndexArrangement(newArr);
+    return newArr;
+  }
+};
+export default deleteTodo;
