+++ src/todo.js
@@ -9,18 +9,10 @@ const IndexArrangement = (arr) => {
   });
 };
 
-const deleteTodo = (arr) => {
-  if (arr.length > 0) {
-    const newArr = arr.filter((todo) => todo.completed === false);
-    IndexArrangement(newArr);
-    return newArr;
-  }
-};
-
 const editTodo = (todo, newDescription) => {
   todo.description = newDescription;
 };
 
 export {
-  deleteTodo, editTodo, Todo,
+  editTodo, Todo, IndexArrangement,
 };
